Fix small bug in setting “ModificationTime” property.
When ‘upload’ is used and when the corresponding file already exists
in Drive but wasn’t originally created by skicka, then skicka adds its
custom properties to the file. Before, we were initializing
“ModificationTime” with the value “0”; in the case where the file on
drive matched the local file, we never updated it.

This in turn caused the modification time to be reported in the year
1969, which in turn caused local files to be md5 checksummed 
unnecessarily on the next upload. (However, after this upload, 
“ModificationTime” would be set properly and all was well.)