Add support for rate-limited uploads.
This is configured via a new bytes-per-second-limit field in the
[upload] section of the .skicka.config file.  If unset (or zero),
then there are no limits.

To do this, skicka now POSTs the upload data directly, rather than
using the Drive API, which in turn allows us to use a rate-limited
io.Reader class to hand out the data in chunks subject to the
user-supplied bandwidth limit.

Addresses the “upload” part of issue #1. Will help a bit with
issue #13 (by making one less copy in memory).