Bugfix: update local modification time from Drive
Before, when we considered downloading a file from Drive because the
Drive version was newer than the local version but it turned out that
the files matched MD5 hashes, we weren’t updating the local
modification time to match the modification time from Drive. This
in turn meant that we would endlessly keep on doing MD5 hashes of the
local file on subsequent download runs.