Print usage when -h, -help, or --help are invoked
The Usage var in the flag package defines the function that is called
when the flag -help or -h is called but no help flag is defined.  The
flag.Usage variable can be reassigned to another function if an
alternate function should be invoked, which is what this commit does.
If "-h", "--h", "-help", "--help" or "help" is typed as the command line
argument, "usage()" is invoked.