Modify download function to show progress based on size
This commit changes the algorithm for how the download sync works.
Specifically, it figures out which files must be downloaded before it
starts trying to download anything.  In this manner, it's possible to
calculate how much the total download will be, in contrast to having the
logic whether something should be downloaded or not within the download
function itself.

The upload functionality will come next.