Add error checking and improve upload progress
This extracts some error-checking code that could have been repeated
into a function that increments download errors and prints a message.

This commit also adds a parameter to the SyncFileUp function so that the
upload progress bar can more accurately display overall progress to the
user.  This aditional variable is an io.Writer.  This writer gets tee'd
all reads from the source file.   In this specific case, the reads from
the source file are what's uploaded to drive.