Set proper download url for Google Doc files
Google Doc files, in the drive.File entity,
have an empty DownloadUrl field as they are not
stored in Drive. However, they do have an
ExportLinks<mime> map field which has a similar
url to download and export Google Doc files.

This change assigns the ExportLinks<mime> field
if the file url was empty after assigning the
DownloadUrl to it.

Should resolve google/skicka#26.