Count the bytes that are uploaded for progress
If the upload fails then the progress must change.  The change depends
on the type of failure.  If the failure is temporary and an attempt will
be made to re-upload the file, then the progress from that first attempt
is subtracted from the total progress.  Otherwise, if no further attempt
will be made, then the total expected byte uploads will be subtracted by
that file's size.