Add the RM command
This commit adds tests and implements the rm command. It attempts to
mimic the implementation of a unix shell's rm program.  The recursive
flag isn't really necessary because google drive API is fine deleting
folders even if there are contents in them, but since the rm program in
the shell fails on it the -r flag must be passed. This is probably for
the best, since if the -s command is passed as well, everything under
the path you've specified in the drive is removed, skipping the trash.

Also, if the program detects the user made an intention to use the 'rm'
command but has failed to provide the right syntax, rm-specific syntax
is provided instead of printing the entire usage.