Start of better support for multiple files w/same name on Drive.
Drive allows multiple files (and folders) to have the same name,
unfortunately. The gdrive module now provides a GetFilesUnderPath
method, which returns an object that represents the mapping from paths
to *drive.Files, while allowing multiple paths to be present.

The du and ls commands now use this interface and are thus able to
deal with this case. (Partially addresses issue #18).

Also some small cleanups in gdrive and its error handling.