Cache metadata of files on Drive locally.
We now keep a cache of information about all of the files on Drive
locally. Thus, at startup time, we only need to ask Drive for any changes
since the last sync, which is a lot faster than fetching all of the file
metadata in a folder hierarchy when doing an upload or download, etc.

With this, operations like “skicka du /“ run in a second or two, which
is a massive improvement.

There were a fair number of changes needed in the gdrive client code,
some stemming from general API cleanups and others from the fact that
the gdrive.File struct now just contains things like the file size and
modification time as public members, rather than requiring a method
call to get those values.

There’s probably some additional opportunity to simplify client code
to take advantage of the fact that queries for information about files
on Drive are now very efficient; this can allow a much more fine-grained
interaction with the gdrive interfaces rather than the current “give me
all of the information about a file hierarchy” model in the upload and
download code.

Fixes issue #2.