Fix upload bug where failures might not be reported to the user.
In one of the error paths in uploadFileContents, it was possible that after
a series of failures (and sleeps with exponential backoff), that an error
would be returned from upload but the function would return no error. This
could cause the user to incorrectly believe that all files had been
successfully uploaded when they hadn't.

This should only have come up during serious issues like network outages in
the middle of an upload.