Fix metadata cache bug where deleted files would show up in /.
If a folder is trashed, we don’t get *drive.Change messages for each
file in the folder indicating that it was deleted. Therefore, we need
to notice if one of the parent folders of a file was trashed while
we’re creating the maps from file and directory names to *Files.