Fix small error in fileNeedsUpload.
Before, when running with -ignore-times, in the case where the file
modification times were the same but the file contents were different,
fileNeedsUpload would return true, even though it was supposed to
return false (do not upload).

In practice, this didn’t matter, since it also returned an error, and
the calling code doesn’t upload the file in this case.